requirements.txt
streamlit
pandas
google-genai
python-dotenv
firebase-admin
plotly
geopy
